Split up QueryManagerConfig
This renames the following property:

  query.shard.max-threads -> task.shard.max-threads

Additionally, a new config property task.info.max-age is split out
from the query.info.max-age property. The task property controls
how often the coordinator must heartbeat tasks on the workers. The
query property controls how often the end user (statement client)
must heartbeat the query on the coordinator.
